refactor: 💡 add validation when update, read and delete Task
add the validation to validate the user should able to get/edit thier
own task only.

+++ core/application/use-cases/delete-task.use-case.ts
@@ -1,17 +1,31 @@
 import { UseCase } from '../../shared/interfaces/use-case.interface';
 import { TaskRepositoryPort } from '../ports/task-repository.port';
 import { TaskNotFoundException } from '@core/shared/exceptions/task-notfound.exception';
+import { UnauthorizedException } from '@core/shared/exceptions/accessdenied.exception';
 import { TaskId } from '@core/domain/value-objects/task-id.vo';
+import { UserId } from '@core/domain/value-objects/user-id.vo';
 
-export class DeleteTaskUseCase implements UseCase<string, void> {
+export interface DeleteTaskUseCaseInput {
+  id: string;
+  userId: UserId;
+}
+
+export class DeleteTaskUseCase
+  implements UseCase<DeleteTaskUseCaseInput, void>
+{
   constructor(private readonly TaskRepository: TaskRepositoryPort) {}
 
-  async execute(id: string): Promise<void> {
-    const taskId = new TaskId(id);
+  async execute(input: DeleteTaskUseCaseInput): Promise<void> {
+    const taskId = new TaskId(input.id);
     const existingTask = await this.TaskRepository.findById(taskId);
 
     if (!existingTask) {
       throw new TaskNotFoundException(taskId);
+    }
+
+    const isOwner = await this.TaskRepository.isOwner(taskId, input.userId);
+    if (!isOwner) {
+      throw new UnauthorizedException();
     }
 
     await this.TaskRepository.delete(taskId);
